Change files for move
