+++ strax/processing/data_reduction.py
@@ -3,8 +3,8 @@ import numpy as np
 import numba
 from enum import IntEnum
 
-from strax.pulse_processing import NOT_APPLICABLE, record_links
-from strax.peak_building import find_peaks
+from strax.processing.pulse_processing import NOT_APPLICABLE, record_links
+from strax.processing.peak_building import find_peaks
 from strax.utils import fully_contained_in
 from strax.dtypes import peak_dtype
